# TrieNode further memory optimizations