--- a/src/Nethermind/Nethermind.Trie.Test/Pruning/TreeStoreTests.cs
+++ b/src/Nethermind/Nethermind.Trie.Test/Pruning/TreeStoreTests.cs
@@ -253,7 +253,7 @@ namespace Nethermind.Trie.Test.Pruning
 
             using TrieStore trieStore = new(memDb, new MemoryLimit(16.MB()), new ConstantInterval(4), _logManager);
 
-            trieStore.CommitNode(0, new NodeCommitInfo(a));
+            trieStore.CommitNode(1, new NodeCommitInfo(a));
             trieStore.FinishBlockCommit(TrieType.State, 0, a);
             trieStore.FinishBlockCommit(TrieType.State, 1, a);
             trieStore.FinishBlockCommit(TrieType.State, 2, a);
@@ -274,7 +274,7 @@ namespace Nethermind.Trie.Test.Pruning
 
             using TrieStore trieStore = new(memDb, new MemoryLimit(16.MB()), No.Persistence, _logManager);
 
-            trieStore.CommitNode(0, new NodeCommitInfo(a));
+            trieStore.CommitNode(1, new NodeCommitInfo(a));
             trieStore.FinishBlockCommit(TrieType.State, 0, a);
             trieStore.FinishBlockCommit(TrieType.State, 1, a);
             trieStore.FinishBlockCommit(TrieType.State, 2, a);
@@ -560,7 +560,7 @@ namespace Nethermind.Trie.Test.Pruning
             WitnessCollector witnessCollector = new WitnessCollector(new MemDb(), LimboLogs.Instance);
             IKeyValueStoreWithBatching store = originalStore.WitnessedBy(witnessCollector);
             using TrieStore trieStore = new(store, new TestPruningStrategy(false), No.Persistence, _logManager);
-            trieStore.CommitNode(0, new NodeCommitInfo(node));
+            trieStore.CommitNode(1, new NodeCommitInfo(node));
             trieStore.FinishBlockCommit(TrieType.State, 0, node);
 
             IReadOnlyTrieStore readOnlyTrieStore = trieStore.AsReadOnly(originalStore);
@@ -636,7 +636,7 @@ namespace Nethermind.Trie.Test.Pruning
             node.ResolveKey(NullTrieNodeResolver.Instance, true);
 
             using TrieStore trieStore = new(new MemDb(), new TestPruningStrategy(pruning), No.Persistence, _logManager);
-            trieStore.CommitNode(0, new NodeCommitInfo(node));
+            trieStore.CommitNode(1, new NodeCommitInfo(node));
             trieStore.FinishBlockCommit(TrieType.State, 0, node);
             var originalNode = trieStore.FindCachedOrUnknown(node.Keccak);