--- a/src/Nethermind/Nethermind.Trie.Test/TrieNodeTests.cs
+++ b/src/Nethermind/Nethermind.Trie.Test/TrieNodeTests.cs
@@ -171,7 +171,8 @@ namespace Nethermind.Trie.Test
             decodedTiniest.ResolveNode(NullTrieNodeResolver.Instance);
 
             Assert.AreEqual(ctx.TiniestLeaf.Value, decodedTiniest.Value, "value");
-            Assert.AreEqual(HexPrefix.ToBytes(ctx.TiniestLeaf.Key!, true), HexPrefix.ToBytes(decodedTiniest.Key!, true), "key");
+            Assert.AreEqual(HexPrefix.ToBytes(ctx.TiniestLeaf.Key!, true), HexPrefix.ToBytes(decodedTiniest.Key!, true),
+                "key");
         }
 
         [Test]
@@ -511,14 +512,14 @@ namespace Nethermind.Trie.Test
         public void Size_of_a_heavy_leaf_is_correct()
         {
             Context ctx = new();
-            Assert.AreEqual(240, ctx.HeavyLeaf.GetMemorySize(false));
+            Assert.AreEqual(232, ctx.HeavyLeaf.GetMemorySize(false));
         }
 
         [Test]
         public void Size_of_a_tiny_leaf_is_correct()
         {
             Context ctx = new();
-            Assert.AreEqual(160, ctx.TiniestLeaf.GetMemorySize(false));
+            Assert.AreEqual(152, ctx.TiniestLeaf.GetMemorySize(false));
         }
 
         [Test]
@@ -532,8 +533,8 @@ namespace Nethermind.Trie.Test
                 node.SetChild(i, ctx.AccountLeaf);
             }
 
-            Assert.AreEqual(3792, node.GetMemorySize(true));
-            Assert.AreEqual(208, node.GetMemorySize(false));
+            Assert.AreEqual(3656, node.GetMemorySize(true));
+            Assert.AreEqual(200, node.GetMemorySize(false));
         }
 
         [Test]
@@ -544,7 +545,7 @@ namespace Nethermind.Trie.Test
             trieNode.Key = new byte[] { 1 };
             trieNode.SetChild(0, ctx.TiniestLeaf);
 
-            Assert.AreEqual(128, trieNode.GetMemorySize(false));
+            Assert.AreEqual(120, trieNode.GetMemorySize(false));
         }
 
         [Test]
@@ -555,22 +556,22 @@ namespace Nethermind.Trie.Test
             trieNode.Key = new byte[] { 1 };
             trieNode.SetChild(0, ctx.TiniestLeaf);
 
-            Assert.AreEqual(288, trieNode.GetMemorySize(true));
-            Assert.AreEqual(128, trieNode.GetMemorySize(false));
+            Assert.AreEqual(272, trieNode.GetMemorySize(true));
+            Assert.AreEqual(120, trieNode.GetMemorySize(false));
         }
 
         [Test]
         public void Size_of_an_unknown_empty_node_is_correct()
         {
             TrieNode trieNode = new(NodeType.Unknown);
-            trieNode.GetMemorySize(false).Should().Be(56);
+            trieNode.GetMemorySize(false).Should().Be(48);
         }
 
         [Test]
         public void Size_of_an_unknown_node_with_keccak_is_correct()
         {
             TrieNode trieNode = new(NodeType.Unknown, Keccak.Zero);
-            trieNode.GetMemorySize(false).Should().Be(136);
+            trieNode.GetMemorySize(false).Should().Be(128);
         }
 
         [Test]
@@ -578,7 +579,7 @@ namespace Nethermind.Trie.Test
         {
             TrieNode trieNode = new(NodeType.Extension);
             trieNode.SetChild(0, null);
-            trieNode.GetMemorySize(false).Should().Be(96);
+            trieNode.GetMemorySize(false).Should().Be(88);
         }
 
         [Test]
@@ -586,7 +587,7 @@ namespace Nethermind.Trie.Test
         {
             TrieNode trieNode = new(NodeType.Branch);
             trieNode.SetChild(0, null);
-            trieNode.GetMemorySize(false).Should().Be(208);
+            trieNode.GetMemorySize(false).Should().Be(200);
         }
 
         [Test]
@@ -594,14 +595,14 @@ namespace Nethermind.Trie.Test
         {
             TrieNode trieNode = new(NodeType.Leaf);
             trieNode.Value = new byte[7];
-            trieNode.GetMemorySize(false).Should().Be(136);
+            trieNode.GetMemorySize(false).Should().Be(128);
         }
 
         [Test]
         public void Size_of_an_unknown_node_with_full_rlp_is_correct()
         {
             TrieNode trieNode = new(NodeType.Unknown, new byte[7]);
-            trieNode.GetMemorySize(false).Should().Be(120);
+            trieNode.GetMemorySize(false).Should().Be(80);
         }
 
         [Test]
@@ -699,7 +700,7 @@ namespace Nethermind.Trie.Test
             trieNode.SetChild(0, child);
 
             trieNode.PrunePersistedRecursively(1);
-            trieNode.GetMemorySize(false).Should().Be(176);
+            trieNode.GetMemorySize(false).Should().Be(168);
         }
 
         [Test]
@@ -712,7 +713,7 @@ namespace Nethermind.Trie.Test
             trieNode.PrunePersistedRecursively(1);
             TrieNode cloned = trieNode.Clone();
 
-            cloned.GetMemorySize(false).Should().Be(176);
+            cloned.GetMemorySize(false).Should().Be(168);
         }
 
         [Test]
@@ -884,14 +885,14 @@ namespace Nethermind.Trie.Test
             leaf1.Value = new byte[111];
             leaf1.ResolveKey(trieStore, false);
             leaf1.Seal();
-            trieStore.CommitNode(0, new NodeCommitInfo(leaf1));
+            trieStore.CommitNode(1, new NodeCommitInfo(leaf1));
 
             TrieNode leaf2 = new(NodeType.Leaf);
             leaf2.Key = Bytes.FromHexString("abd");
             leaf2.Value = new byte[222];
             leaf2.ResolveKey(trieStore, false);
             leaf2.Seal();
-            trieStore.CommitNode(0, new NodeCommitInfo(leaf2));
+            trieStore.CommitNode(1, new NodeCommitInfo(leaf2));
 
             TrieNode trieNode = new(NodeType.Branch);
             trieNode.SetChild(1, leaf1);