This PR pushes further with memory optimizations of `TrieNode`, that were initiated in #5259. `TrieNode` is an object that is cached in memory and usually has ~350,000 objects of this type allocated. This PR cuts another 16 bytes from the class.

## Changes

- make `storageRoot` a part of data so that only leaves allocate it (reduce footprint by 8 bytes for non-leaves)
- remove `FullRlp` reference and delegate to `_rlpStream.Data`
- amend `GetMemorySize` as it was previously (**thorough review required**) double counting the data (once under FullRlp, once under rlpstream)

## Benchmarks

```bash
BenchmarkDotNet=v0.13.4, OS=Windows 11 (10.0.22621.1105)
12th Gen Intel Core i9-12900HK, 1 CPU, 20 logical and 14 physical cores
.NET SDK=7.0.100
  [Host]     : .NET 7.0.0 (7.0.22.51805), X64 RyuJIT AVX2
  DefaultJob : .NET 7.0.0 (7.0.22.51805), X64 RyuJIT AVX2
```

|        Method |     Mean |     Error |    StdDev |   Gen0 | Allocated |
|-------------- |---------:|----------:|----------:|-------:|----------:|
| PropertiesSet (master) | 7.509 ns | 0.0954 ns | 0.0846 ns | 0.0064 |     64 B  |
| PropertiesSet ([85ec8ef](https://github.com/NethermindEth/nethermind/pull/5262/commits/85ec8ef97c95e39dc13f3b85281ac287e6a07772))  - storageRoot| 7.223 ns | 0.1418 ns | 0.2407 ns | 0.0045 |      56 B |
| PropertiesSet ([e3049e0](https://github.com/NethermindEth/nethermind/pull/5262/commits/e3049e02732a8dd8a26c07116fee860d06a970a1)) - FullRlp ref | 7.089 ns | 0.1305 ns | 0.1221 ns | 0.0038 |      48 B |

## Types of changes

#### What types of changes does your code introduce?

- [ ] Bugfix (a non-breaking change that fixes an issue)
- [ ] New feature (a non-breaking change that adds functionality)
- [ ] Breaking change (a change that causes existing functionality not to work as expected)
- [x] Optimization
- [ ] Refactoring
- [ ] Documentation update
- [ ] Build-related changes
- [ ] Other: _Description_

## Testing

#### Requires testing

- [] Yes
- [x] No

#### If yes, did you write tests?

- [ ] Yes
- [x] No

No functional changes

## Remarks

Low level optimizations make code a bit ugly. Still, something that is allocated ~350,000 times can be a bit uglier.